fix(localPush): Fix local push notification does not show. (#153)
